Move how to play popups out of GameplayOptions
